+++ src/hydrogen/core/duct_template.clj
@@ -24,7 +24,7 @@
            [re-frame "0.10.6"]
            [reagent "0.8.1"]
            [secretary "1.2.3"]
-           [hydrogen/module "0.1.4"]]
+           [hydrogen/module.core "0.1.4"]]
    :dev-deps '[[day8.re-frame/re-frame-10x "0.3.7"]]
    :templates {;; Client
                "src/{{dirs}}/client.cljs" (resource "core/cljs/client.cljs")