+++ src/hydrogen/session/keycloak/duct_template.clj
@@ -9,11 +9,13 @@
   {:vars {:hydrogen-session? true
           :hydrogen-session-keycloak? true}
    :deps '[[duct/middleware.buddy "0.1.0"]
-           [magnet/buddy-auth.jwt-oidc "0.5.0"]]
+           [magnet/buddy-auth.jwt-oidc "0.5.0"]
+           [hydrogen/module.session.keycloak "0.1.4"]]
    :templates {;; Client
-               "src/{{dirs}}/client/landing.cljs" (resource "keycloak/cljs/landing.cljs")
-               "src/{{dirs}}/client/session.cljs" (resource "keycloak/cljs/session.cljs")
+               "src/{{dirs}}/client/cookie.cljs" (resource "session/keycloak/cljs/cookie.cljs")
+               "src/{{dirs}}/client/landing.cljs" (resource "session/keycloak/cljs/landing.cljs")
+               "src/{{dirs}}/client/session.cljs" (resource "session/keycloak/cljs/session.cljs")
                ;; Resources
-               "resources/{{dirs}}/public/css/auth.scss" (resource "keycloak/resources/css/auth.scss")
-               "resources/{{dirs}}/public/css/landing.scss" (resource "keycloak/resources/css/landing.scss")}
+               "resources/{{dirs}}/public/css/auth.scss" (resource "session/keycloak/resources/css/auth.scss")
+               "resources/{{dirs}}/public/css/landing.scss" (resource "session/keycloak/resources/css/landing.scss")}
    :modules {:hydrogen.module/session.keycloak {:add-example-api? true}}})